+++ README.md
@@ -49,12 +49,60 @@ the `--profile <name>` flag when interacting with the API, or to specfiy the pro
 
 The `./deploy` file is created in the root directory of the git repo and follows the following format:
 
-```
+```JSON
 {
     "profile": "name of AWS Account profile",
     "tags": [{
         "Key": "Optional Tags",
         "Value": "That should be applied to Resources (Useful for billing)"
     }]
+}
+```
+
+### Watching for Artifacts
+
+By default, if a `Dockerfile` is found in the project root, the ECR will be queried before deploy to ensure
+the image has been built. IE: a git repo named `my-project` would look for an image called `my-project:<Git Sha>`.
+
+If you are building multiple docker images, or want to disable this feature, the following options are avaliable
+via the `.deploy` file.
+
+**Disable ECR Check**
+
+```JSON
+{
+    "artifacts": {
+        "docker": false
+    }
+}
+```
+**Custom Pre/Postfix**
+
+```JSON
+{
+    "artifacts": {
+        "docker": "custom-ecr:{{gitsha}}"
+    }
+}
+```
+
+```JSON
+{
+    "artifacts": {
+        "docker": "{{project}}:prefix-{{gitsha}}-postfix"
+    }
+}
+```
+
+**Multiple Images**
+
+```JSON
+{
+    "artifacts": {
+        "docker": [
+            "{{project}}:backend-{{gitsha}}",
+            "{{project}}:frontend-{{gitsha}}"
+        ]
+    }
 }
 ```
